# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

/**
 * The tone for the reply
 */
export type GenerateEmailReplyBodyTone =
  (typeof GenerateEmailReplyBodyTone)[keyof typeof GenerateEmailReplyBodyTone];

export const GenerateEmailReplyBodyTone = {
  professional: "professional",
  friendly: "friendly",
  firm: "firm",
  empathetic: "empathetic",
} as const;

/**
 * The use case context
 */
export type GenerateEmailReplyBodyUseCase =
  (typeof GenerateEmailReplyBodyUseCase)[keyof typeof GenerateEmailReplyBodyUseCase];

export const GenerateEmailReplyBodyUseCase = {
  "general-work": "general-work",
  "customer-support": "customer-support",
  recruiter: "recruiter",
  client: "client",
} as const;

/**
 * The desired reply length
 */
export type GenerateEmailReplyBodyLength =
  (typeof GenerateEmailReplyBodyLength)[keyof typeof GenerateEmailReplyBodyLength];

export const GenerateEmailReplyBodyLength = {
  short: "short",
  medium: "medium",
  detailed: "detailed",
} as const;

export interface GenerateEmailReplyBody {
  /** The original email content to reply to */
  emailContent: string;
  /** The tone for the reply */
  tone: GenerateEmailReplyBodyTone;
  /** The use case context */
  useCase: GenerateEmailReplyBodyUseCase;
  /** The desired reply length */
  length: GenerateEmailReplyBodyLength;
}

/**
 * Risk level of the email
 */
export type GenerateEmailReplyResponseRiskLevel =
  (typeof GenerateEmailReplyResponseRiskLevel)[keyof typeof GenerateEmailReplyResponseRiskLevel];

export const GenerateEmailReplyResponseRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface GenerateEmailReplyResponse {
  /** The primary recommended reply */
  bestReply: string;
  /** Two alternative reply approaches */
  alternativeReplies: string[];
  /** Mobile-friendly 1-2 sentence version */
  shortReply: string;
  /** Suggested subject line */
  suggestedSubject: string;
  /** Detected sentiment of the incoming email */
  sentiment: string;
  /** Detected intent of the sender */
  intent: string;
  /** Risk level of the email */
  riskLevel: GenerateEmailReplyResponseRiskLevel;
  /** Explanation of why the email is flagged as risky */
  riskReason?: string;
}