# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Generate AI-powered email replies with tone control, sentiment analysis, and multiple reply variants
 * @summary Generate email replies
 */
export const GenerateEmailReplyBody = zod.object({
  emailContent: zod.string().describe("The original email content to reply to"),
  tone: zod
    .enum(["professional", "friendly", "firm", "empathetic"])
    .describe("The tone for the reply"),
  useCase: zod
    .enum(["general-work", "customer-support", "recruiter", "client"])
    .describe("The use case context"),
  length: zod
    .enum(["short", "medium", "detailed"])
    .describe("The desired reply length"),
});

export const GenerateEmailReplyResponse = zod.object({
  bestReply: zod.string().describe("The primary recommended reply"),
  alternativeReplies: zod
    .array(zod.string())
    .describe("Two alternative reply approaches"),
  shortReply: zod.string().describe("Mobile-friendly 1-2 sentence version"),
  suggestedSubject: zod.string().describe("Suggested subject line"),
  sentiment: zod.string().describe("Detected sentiment of the incoming email"),
  intent: zod.string().describe("Detected intent of the sender"),
  riskLevel: zod
    .enum(["low", "medium", "high"])
    .describe("Risk level of the email"),
  riskReason: zod
    .string()
    .optional()
    .describe("Explanation of why the email is flagged as risky"),
});