# lib/api-zod/src/generated/types/generateEmailReplyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateEmailReplyBodyLength } from "./generateEmailReplyBodyLength";
import type { GenerateEmailReplyBodyTone } from "./generateEmailReplyBodyTone";
import type { GenerateEmailReplyBodyUseCase } from "./generateEmailReplyBodyUseCase";

export interface GenerateEmailReplyBody {
  /** The original email content to reply to */
  emailContent: string;
  /** The tone for the reply */
  tone: GenerateEmailReplyBodyTone;
  /** The use case context */
  useCase: GenerateEmailReplyBodyUseCase;
  /** The desired reply length */
  length: GenerateEmailReplyBodyLength;
}