# lib/api-zod/src/generated/types/generateEmailReplyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateEmailReplyResponseRiskLevel } from "./generateEmailReplyResponseRiskLevel";

export interface GenerateEmailReplyResponse {
  /** The primary recommended reply */
  bestReply: string;
  /** Two alternative reply approaches */
  alternativeReplies: string[];
  /** Mobile-friendly 1-2 sentence version */
  shortReply: string;
  /** Suggested subject line */
  suggestedSubject: string;
  /** Detected sentiment of the incoming email */
  sentiment: string;
  /** Detected intent of the sender */
  intent: string;
  /** Risk level of the email */
  riskLevel: GenerateEmailReplyResponseRiskLevel;
  /** Explanation of why the email is flagged as risky */
  riskReason?: string;
}